refactor: extract decision-resolution helpers and co-locate the plan-dispatcher setter
Move resolve_decision_reason and record_chosen_option out of the approvals _shared grab-bag into a cohesive _decision_resolution sibling, and relocate attach_plan_dispatcher into ProposeActMixin beside the _plan_dispatcher field it sets and the act path reads. Un-export the now-internal decisionOptionsOf helper and accept the writeup/compensatable domain terms in the vale vocabulary.
