+++ src/synthorg/meta/chief_of_staff/propose.py
@@ -194,16 +194,6 @@ class ChiefOfStaffProposer(ProposeActMixin):
         # other caller can address the id before it's assigned.
         self._locks = ConversationLockRegistry()
 
-    def attach_plan_dispatcher(self, dispatcher: ConversationalPlanDispatcher) -> None:
-        """Attach the conversational plan dispatcher (late-bind seam).
-
-        The dispatcher drives an accepted work brief into the plan-review
-        spine (provision project, intake the objective, background the
-        decompose+park). Wired by the startup hook once the work pipeline
-        and background-dispatch port are available.
-        """
-        self._plan_dispatcher = dispatcher
-
     async def _routing_enabled(self) -> bool:
         """Resolve the per-turn concern-routing gate live.
